% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{.rbfmvar_irf_array}
\alias{.rbfmvar_irf_array}
\title{Moving Average Coefficients of a Levels VAR}
\usage{
.rbfmvar_irf_array(A, Sigma_e, horizon, ortho)
}
\description{
Moving Average Coefficients of a Levels VAR
}
\keyword{internal}
