% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{.rbfmvar_levels}
\alias{.rbfmvar_levels}
\title{Levels VAR Coefficients Implied by the RBFM-VAR Estimates}
\usage{
.rbfmvar_levels(F, n, p)
}
\arguments{
\item{F}{Coefficient matrix (regressors x n) in the order
\eqn{\Gamma_1, \dots, \Gamma_{p-2}, \Pi_1, \Pi_2}.}

\item{n}{Number of variables.}

\item{p}{Lag order.}
}
\value{
List of n x n matrices \eqn{A_1, \dots, A_P}, \eqn{P = \max(p, 2)}.
}
\description{
Solves \eqn{(1-L)^2 y_t = \sum_j \Gamma_j L^j (1-L)^2 y_t + \Pi_1 L (1-L)
y_t + \Pi_2 L y_t + e_t} for \eqn{y_t = \sum_k A_k y_{t-k} + e_t}.
}
\keyword{internal}
