% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.rbfmvar}
\alias{fitted.rbfmvar}
\title{Extract Fitted Values from rbfmvar Object}
\usage{
\method{fitted}{rbfmvar}(object, ...)
}
\arguments{
\item{object}{An \code{rbfmvar} object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Matrix of fitted values.
}
\description{
Extract Fitted Values from rbfmvar Object
}
