% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Out-of-Sample Forecasting}
\usage{
forecast(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
The generic function \code{forecast} computes forecasts from time series models.
}
