% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{format_kernel}
\alias{format_kernel}
\title{Format Kernel Name for Display}
\usage{
format_kernel(kernel)
}
\arguments{
\item{kernel}{Kernel name.}
}
\value{
Formatted kernel name.
}
\description{
Format Kernel Name for Display
}
\keyword{internal}
