% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{get_kernel_constant}
\alias{get_kernel_constant}
\title{Get Kernel Rate Constant}
\usage{
get_kernel_constant(kernel)
}
\arguments{
\item{kernel}{Character string specifying the kernel type.}
}
\value{
Rate constant.
}
\description{
Returns the rate constant c_gamma used in Andrews (1991)
  automatic bandwidth selection.
}
\keyword{internal}
