% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{get_kernel_function}
\alias{get_kernel_function}
\title{Get Kernel Function by Name}
\usage{
get_kernel_function(kernel)
}
\arguments{
\item{kernel}{Character string specifying the kernel type.}
}
\value{
Kernel function.
}
\description{
Get Kernel Function by Name
}
\keyword{internal}
