% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granger.R
\name{granger_matrix}
\alias{granger_matrix}
\title{Granger Causality Matrix}
\usage{
granger_matrix(object)
}
\arguments{
\item{object}{An \code{rbfmvar} object.}
}
\value{
A matrix of p-values for all pairwise Granger causality tests.
  Row i, column j contains the p-value for "variable j causes variable i".
}
\description{
Computes pairwise Granger causality tests for all variable pairs.
}
\examples{
\donttest{
# Generate example data
set.seed(123)
n <- 100
mydata <- data.frame(x = cumsum(rnorm(n)), y = cumsum(rnorm(n)))
fit <- rbfmvar(mydata, lags = 2)
granger_matrix(fit)
}

}
