% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_selection.R
\name{ic_table}
\alias{ic_table}
\title{Get Information Criteria Table}
\usage{
ic_table(object, max_lags = 8)
}
\arguments{
\item{object}{An \code{rbfmvar} object.}

\item{max_lags}{Maximum lag order to evaluate.}
}
\value{
A data frame with AIC, BIC, and HQ values.
}
\description{
Returns a table of information criteria values for different
  lag orders.
}
\examples{
\donttest{
# Generate example data
set.seed(123)
n <- 100
mydata <- data.frame(x = cumsum(rnorm(n)), y = cumsum(rnorm(n)))
fit <- rbfmvar(mydata, lags = 2)
ic_table(fit, max_lags = 6)
}

}
