% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel_bartlett}
\alias{kernel_bartlett}
\title{Bartlett (Newey-West) Kernel}
\usage{
kernel_bartlett(x)
}
\arguments{
\item{x}{Numeric scalar or vector of kernel arguments.}
}
\value{
Kernel weights.
}
\description{
Bartlett (Newey-West) Kernel
}
\keyword{internal}
