% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel_parzen}
\alias{kernel_parzen}
\title{Parzen Kernel}
\usage{
kernel_parzen(x)
}
\arguments{
\item{x}{Numeric scalar or vector of kernel arguments.}
}
\value{
Kernel weights.
}
\description{
Parzen Kernel
}
\keyword{internal}
