% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel_qs}
\alias{kernel_qs}
\title{Quadratic Spectral (QS) Kernel}
\usage{
kernel_qs(x)
}
\arguments{
\item{x}{Numeric scalar or vector of kernel arguments.}
}
\value{
Kernel weights.
}
\description{
Quadratic Spectral (QS) Kernel
}
\keyword{internal}
