% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{plot.rbfmvar_forecast}
\alias{plot.rbfmvar_forecast}
\title{Plot Method for rbfmvar_forecast Objects}
\usage{
\method{plot}{rbfmvar_forecast}(x, ...)
}
\arguments{
\item{x}{An \code{rbfmvar_forecast} object.}

\item{...}{Additional arguments passed to \code{plot}.}
}
\value{
No return value, called for side effects (produces a plot).
}
\description{
Plots forecasts from an RBFM-VAR model.
}
