% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.rbfmvar}
\alias{residuals.rbfmvar}
\title{Extract Residuals from rbfmvar Object}
\usage{
\method{residuals}{rbfmvar}(object, ...)
}
\arguments{
\item{object}{An \code{rbfmvar} object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Matrix of residuals.
}
\description{
Extract Residuals from rbfmvar Object
}
