% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrv.R
\name{select_bandwidth_andrews}
\alias{select_bandwidth_andrews}
\title{Andrews (1991) Automatic Bandwidth Selection}
\usage{
select_bandwidth_andrews(v, kernel = "bartlett")
}
\arguments{
\item{v}{Matrix of residuals (T x n).}

\item{kernel}{Character string specifying the kernel type.}
}
\value{
Optimal bandwidth.
}
\description{
Implements the data-dependent automatic bandwidth selection
  procedure of Andrews (1991).
}
\references{
Andrews, D. W. K. (1991). Heteroskedasticity and Autocorrelation Consistent
Covariance Matrix Estimation. \emph{Econometrica}, 59(3), 817-858.
\doi{10.2307/2938229}
}
\keyword{internal}
