% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.rbfmvar}
\alias{vcov.rbfmvar}
\title{Extract Variance-Covariance Matrix from rbfmvar Object}
\usage{
\method{vcov}{rbfmvar}(object, ...)
}
\arguments{
\item{object}{An \code{rbfmvar} object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Error covariance matrix.
}
\description{
Extract Variance-Covariance Matrix from rbfmvar Object
}
