% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrv.R
\name{estimate_onesided_lrv}
\alias{estimate_onesided_lrv}
\title{Estimate One-Sided Long-Run Covariance}
\usage{
estimate_onesided_lrv(e, v, kernel = "bartlett", bandwidth)
}
\arguments{
\item{e}{Matrix of residuals e (T x n1).}

\item{v}{Matrix of residuals v (T x n2).}

\item{kernel}{Character string specifying the kernel type.}

\item{bandwidth}{Bandwidth parameter.}
}
\value{
One-sided long-run covariance matrix (n1 x n2).
}
\description{
Estimates the one-sided long-run covariance matrix.
}
\keyword{internal}
