% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_selection.R
\name{estimate_var_ols}
\alias{estimate_var_ols}
\title{Simple VAR OLS Estimation}
\usage{
estimate_var_ols(data, p)
}
\arguments{
\item{data}{Data matrix.}

\item{p}{Lag order.}
}
\value{
List with residual covariance matrix.
}
\description{
Estimates a standard VAR model via OLS for lag selection.
}
\keyword{internal}
