% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Out-of-Sample Forecasting}
\usage{
forecast(object, ...)
}
\arguments{
\item{object}{A model object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Depends on the method dispatched. See \code{\link{forecast.rbfmvar}}
  for the RBFM-VAR method, which returns an object of class
  \code{"rbfmvar_forecast"}.
}
\description{
The generic function \code{forecast} computes forecasts from time series models.
}
