% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.rbfmvar}
\alias{print.rbfmvar}
\title{Print Method for rbfmvar Objects}
\usage{
\method{print}{rbfmvar}(x, ...)
}
\arguments{
\item{x}{An \code{rbfmvar} object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Prints a summary of an RBFM-VAR estimation.
}
