% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbfmvar.R
\name{rbfmvar_estimate}
\alias{rbfmvar_estimate}
\title{Core RBFM-VAR Estimation}
\usage{
rbfmvar_estimate(Y, p, kernel, bandwidth)
}
\arguments{
\item{Y}{Data matrix (T x n).}

\item{p}{Lag order.}

\item{kernel}{Kernel type.}

\item{bandwidth}{Bandwidth (-1 for automatic).}
}
\value{
List of estimation results.
}
\description{
Core RBFM-VAR Estimation
}
\keyword{internal}
